- correction bug (impossible de creer ou uploader) suite a la mise du nouveau systeme de securite check_file_allowed() - rafraichir la liste apres action - suppression de die pour rester user friendly meme avec les sans genes ;)
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@2922 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ _amelioration_admin_/skel_editor/exec/skeleditor.php
@@ -232,13 +232,11 @@ function skel_parser_param($str,$output='') {
   return $output; 
 }
 
-// security: avoid access to other directories
+// security
 function check_file_allowed($file,$files_editable,$new = false) {
-	if (in_array($file,$files_editable))
-		return true;
-	else if ($new){
-		if (in_array(dirname($file),array_map('dirname',$files_editable)))
-			return true;
+	if (in_array($file,$files_editable))	return true;  // known file
+	 else if ($new){ // new file ?	   
+		 if (in_array(dirname($file),array_map('dirname',$files_editable)))	return true; // known directory    
 	}
 	return false;
 }
@@ -297,7 +295,7 @@ function exec_skeleditor(){
   include_spip("inc/presentation");
   global $spip_lang_right;
   $img_extension = array("jpg","png","gif","ico","bmp");
-  $listed_extension = array("htm","html","xml","svg","php","php3","php4","py","sh","sql","css","rdf","txt","nfo","log","js","as");
+  $listed_extension = array("htm","html","xml","svg","php","php3","php4","py","sh","sql","css","rdf","txt","nfo","log","js","as","csv");
 
   // check rights
   global $connect_statut;
@@ -325,17 +323,15 @@ function exec_skeleditor(){
 	// Action ? 
 	// ---------------------------------------------------------------------------
 	$log = "";
+	$safe_flag = false;
 	// POST request ?
 	if (isset($_POST['editor'])) {      // save file ?
 	     $editor = $_POST['editor'];
 	     $editor = str_replace("&lt;/textarea","</textarea",$editor); // exception: textarea closing tag	     
 	     if (isset($_GET['f'])){
-	     	$file_name = $_GET['f'];
-		    if (!check_file_allowed($file_name,$files_editable))
-		     	die('faut pas se gener');    // security	     
-	     }
-	     else $file_name = "";
-	     if (is_writable($file_name)) {
+	     	 $file_name = $_GET['f'];		         
+	     } else $file_name = "";
+	     if (is_writable($file_name) && check_file_allowed($file_name,$files_editable)) {
              if (!$handle = fopen($file_name, 'w')) {
                  $log = "<span style='color:red'>"._T("skeleditor:erreur_ouverture_fichier")."</span>";
              } else if (fwrite($handle, $editor) === FALSE) {
@@ -355,68 +351,73 @@ function exec_skeleditor(){
       $tmp_name = $_FILES['upf']['tmp_name'];
       if (isset($_POST['target'])) {      
                $target = ($_POST['target'])."/".$_FILES['upf']['name'];    // security
-					     if (!check_file_allowed($target,$files_editable,true))
-					     	die('faut pas se gener');    // security	     
-               $_GET['f'] = $target;
-               $_GET['action'] = 'preview';
-               if (file_exists($target)) {
-                  $log = "<span style='color:red'>"._T('skeleditor:erreur_overwrite')."</span>";
+					     if (check_file_allowed($target,$files_editable,true)) {     // security
+					           $_GET['f'] = $target;					           
+                     $_GET['action'] = 'preview';
+                     if (file_exists($target)) {
+                        $log = "<span style='color:red'>"._T('skeleditor:erreur_overwrite')."</span>";
+                     } else {
+                        $ok = @copy($tmp_name, $target);                 
+                        if (!$ok) $ok = @move_uploaded_file($tmp_name, $target);
+                        if (!$ok) $log = "<span style='color:red'>"._T('skeleditor:erreur_droits')."</span>";
+                             else $log = "<span style='color:green'>"._T('skeleditor:fichier_upload_ok')."</span>";  
+                     } 
                } else {
-                  $ok = @copy($tmp_name, $target);                 
-                  if (!$ok) $ok = @move_uploaded_file($tmp_name, $target);
-                  if (!$ok) $log = "<span style='color:red'>"._T('skeleditor:erreur_droits')."</span>";
-                       else $log = "<span style='color:green'>"._T('skeleditor:fichier_upload_ok')."</span>";                         
-                       
-               } 
+                     $log = "<span style='color:red'>"._T('skeleditor:erreur_sansgene')."</span>";
+               }
+					     	             
+
       }     
   } 
    
   // GET request ?
   $action = "";
 	if (isset($_GET['f'])) {
-	    $file_name = $_GET['f']; 
-	    if (!check_file_allowed($file_name,$files_editable))
-	     	die('faut pas se gener');    // security	     
-      if (isset($_GET['target'])) {
-          $target = $_GET['target'];
-          $file_name =  $target."/".$file_name;
-			    if (!check_file_allowed($file_name,$files_editable,true))
-			     	die('faut pas se gener');    // security	     
-      }                       	    
-                  
-      if (isset($_GET['action'])) {
+	    $file_name = $_GET['f'];
+	    
+	    // check security first	    
+	    if (isset($_GET['target'])) {      // exception: new file
+	         $target = $_GET['target'];
+           $file_name =  $target."/".$file_name;
+           $safe_flag = check_file_allowed($file_name,$files_editable,true); 
+      } else if (isset($_FILES['upf'])) { // exception: upload file
+           $safe_flag = check_file_allowed($file_name,$files_editable,true); 
+      } else {
+           $safe_flag = check_file_allowed($file_name,$files_editable);
+      }	                          
+	    
+	    if (isset($_GET['action']) && $safe_flag) { // any action on file ?
           $action = $_GET['action'];          
-          if ($action=="delete") {                // delete the file            
+          if ($action=="delete") {                // delete the file 
             @unlink($file_name);          
-          } else if ($action=="download") {       // download the file              
-              if ($file_tmp = @file("$file_name")) {
-                $file_name_nopath = basename($file_name);
-                $file_str = implode ('',$file_tmp);
-                //header("Content-type: text/plain"); // text/plain or binary .... 
-                header("Content-Disposition: attachment; filename=\"$file_name_nopath\"");
-                echo $file_str; 
-                exit; 
-              }
-          } else if ($action=="new") {            // add new file
-            if (isset($_GET['target'])) {               
-               // FIXME: check if extension available ?                   
-               if (is_file($file_name)) { // security 3: ovewrite ?
-                $log = "<span style='color:red'>"._T("skeleditor:erreur_overwrite")."</span>";
-               } else {
-                 if (!$handle = fopen($file_name, 'w')) {
-                     $log = "<span style='color:red'>"._T("skeleditor:erreur_droits")."</span>";
-                 } else if (fwrite($handle, "...") === FALSE) {
-                     $log = "<span style='color:red'>"._T("skeleditor:erreur_droits")."</span>";           
-                 } else {
-                     $log = "<span style='color:green'>"._T("skeleditor:fichier_sauvegarde_date").date('H:i')."</span>";
-                     fclose($handle);
-                 }
-                }
+          } else if ($action=="download") {       // download the file 
+            if ($file_tmp = @file("$file_name")) {
+                  $file_name_nopath = basename($file_name);
+                  $file_str = implode ('',$file_tmp);
+                  //header("Content-type: text/plain"); // text/plain or binary .... 
+                  header("Content-Disposition: attachment; filename=\"$file_name_nopath\"");
+                  echo $file_str; 
+                  exit; 
             }
-           
-             
-              
+          } else if ($action=="new") {            // add new file
+            if (isset($_GET['target'])) {                 
+			          // FIXME: check if allowed extension ?                   
+                if (is_file($file_name)) {  // security : ovewrite ?
+                      $log = "<span style='color:red'>"._T("skeleditor:erreur_overwrite")."</span>";
+                } else {
+                       if (!$handle = fopen($file_name, 'w')) {
+                           $log = "<span style='color:red'>"._T("skeleditor:erreur_droits")."</span>";
+                       } else if (fwrite($handle, "...") === FALSE) {
+                           $log = "<span style='color:red'>"._T("skeleditor:erreur_droits")."</span>";           
+                       } else {
+                           $log = "<span style='color:green'>"._T("skeleditor:fichier_sauvegarde_date").date('H:i')."</span>";
+                           fclose($handle);                     
+                       }
+                 }
+            }
           }
+          // refresh file list after action
+          $files_editable = parse_path($dossier_squelettes,array_merge($listed_extension,$img_extension));
       } 
   } else {
       $file_name = "";  
@@ -443,36 +444,40 @@ function exec_skeleditor(){
 	debut_droite();
 
 	// something to do ?	
-	if ($file_name!="") {   
-       echo "<div>"._T("skeleditor:fichier")."<strong>$file_name</strong> $log</div>\n"; // add extra infos on file:  size ? date ? ...
-       if ($action=="delete") {
-         echo "<p style='color:green'>"._T("skeleditor:fichier_efface_ok")."</p>\n";
-       } else { 
-           // tools bar
-           echo "<div id='skel_toolbar' style='width:100%;text-align:right;'>\n";
-           echo "<img src='"._DIR_PLUGIN_SKELEDITOR."/img_pack/action_dl.png' alt='download' /><a href=\"?exec=skeleditor&amp;f=".urlencode($file_name)."&amp;action=download\">"._T("skeleditor:telecharger")."</a>";
-           echo "<img src='"._DIR_PLUGIN_SKELEDITOR."/img_pack/action_del.png' alt='delete' /><a href=\"?exec=skeleditor&amp;f=".urlencode($file_name)."&amp;action=delete\" onclick=\"javascript:return confirm('"._T("skeleditor:effacer_confirme")."');\">"._T("skeleditor:effacer")."</a>";
-           echo "</div>\n";
-           // img or text ?
-           $extension =  strtolower(substr($file_name, strrpos($file_name,".")+1)); 
-           if (in_array($extension,$img_extension)) {     // display file as img
-              echo "<div style='border:1px solid #333;padding:20px;background:#eee'><img src='$file_name' alt='picture' /></div>\n";
-           } else {  // edit file as text  
-              if ($file_tmp = @file("$file_name")) {
-                  $file_str = implode ('',$file_tmp);                  
-                  if ($extension=='html') echo  skel_parser($file_str); // experimental         	        
-                  $file_str = str_replace("</textarea","&lt;/textarea",$file_str); // exception: textarea closing tag                                
-                  //echo "<form method='post' action='?exec=skeleditor&amp;retour=skeleditor&amp;f=".urlencode($file_name)."'>\n";
-									echo generer_url_post_ecrire('skeleditor',"retour=skeleditor&f=".urlencode($file_name));
-                  echo "<textarea name='editor' cols='80' rows='50'>$file_str</textarea>\n";               
-									echo "<div style='text-align:$spip_lang_right'><input type='submit' name='action' value='"._T("skeleditor:sauver")."' class='fondo'></div>";
-        	        echo "</form>\n";       	        
-
-              } else {
-                  echo "<p>"._T("skeleditor:erreur_ouvert_ecrit")."</p>\n";
-              }   
-           }
-       }
+	if ($file_name!="") { 
+       if ($safe_flag) {         
+         echo "<div>"._T("skeleditor:fichier")."<strong>$file_name</strong> $log</div>\n"; // add extra infos on file:  size ? date ? ...
+         if ($action=="delete") {
+           echo "<p style='color:green'>"._T("skeleditor:fichier_efface_ok")."</p>\n";
+         } else { 
+             // tools bar
+             echo "<div id='skel_toolbar' style='width:100%;text-align:right;'>\n";
+             echo "<img src='"._DIR_PLUGIN_SKELEDITOR."/img_pack/action_dl.png' alt='download' /><a href=\"?exec=skeleditor&amp;f=".urlencode($file_name)."&amp;action=download\">"._T("skeleditor:telecharger")."</a>";
+             echo "<img src='"._DIR_PLUGIN_SKELEDITOR."/img_pack/action_del.png' alt='delete' /><a href=\"?exec=skeleditor&amp;f=".urlencode($file_name)."&amp;action=delete\" onclick=\"javascript:return confirm('"._T("skeleditor:effacer_confirme")."');\">"._T("skeleditor:effacer")."</a>";
+             echo "</div>\n";
+             // img or text ?
+             $extension =  strtolower(substr($file_name, strrpos($file_name,".")+1)); 
+             if (in_array($extension,$img_extension)) {     // display file as img
+                echo "<div style='border:1px solid #333;padding:20px;background:#eee'><img src='$file_name' alt='picture' /></div>\n";
+             } else {  // edit file as text  
+                if ($file_tmp = @file("$file_name")) {
+                    $file_str = implode ('',$file_tmp);                  
+                    if ($extension=='html') echo  skel_parser($file_str); // experimental         	        
+                    $file_str = str_replace("</textarea","&lt;/textarea",$file_str); // exception: textarea closing tag                                
+                    //echo "<form method='post' action='?exec=skeleditor&amp;retour=skeleditor&amp;f=".urlencode($file_name)."'>\n";
+  									echo generer_url_post_ecrire('skeleditor',"retour=skeleditor&f=".urlencode($file_name));
+                    echo "<textarea name='editor' cols='80' rows='50'>$file_str</textarea>\n";               
+  									echo "<div style='text-align:$spip_lang_right'><input type='submit' name='action' value='"._T("skeleditor:sauver")."' class='fondo'></div>";
+          	        echo "</form>\n";       	        
+  
+                } else {
+                    echo "<p style='color:red'>"._T("skeleditor:erreur_ouvert_ecrit")."</p>\n";
+                }   
+             }
+         }
+      } else { // security failure
+        echo "<div style='color:red'>"._T('skeleditor:erreur_sansgene')."</div>\n";      
+      }
   } else {
       echo "<p>"._T("skeleditor:fichier_choix")."</p>\n";
   }
